[feature] - update mta schema struct
